+++ app/src/main/java/com/dmitry_simakov/queue/fragments/calculaton/model_v/MMV_CalculationFragment.java
@@ -11,19 +11,19 @@ import com.dmitry_simakov.queue.fragments.DoubleBarGraphFragment;
 
 public class MMV_CalculationFragment extends Model_V_CalculationFragment {
     
-    protected float[] W_Values = new float[6];
+    protected double[] W_Values = new double[6];
     public static final String W_VALUES = "W_VALUES";
     
     @Override
     protected void getSavedInstanceStates(Bundle savedInstanceState) {
         super.getSavedInstanceStates(savedInstanceState);
-        W_Values = savedInstanceState.getFloatArray(W_VALUES);
+        W_Values = savedInstanceState.getDoubleArray(W_VALUES);
     }
     
     @Override
     public void onSaveInstanceState(Bundle savedInstanceState) {
         super.onSaveInstanceState(savedInstanceState);
-        savedInstanceState.putFloatArray(W_VALUES, W_Values);
+        savedInstanceState.putDoubleArray(W_VALUES, W_Values);
     }
     
     @Override
@@ -41,8 +41,8 @@ public class MMV_CalculationFragment extends Model_V_CalculationFragment {
     @Override
     protected void getP_Values() {
         Log.d("LOG", "MMV_CalculationFragment: getP_Values");
-        float Pk_Values[] = model.getP();
-        P_Values = new float[V];
+        double Pk_Values[] = model.getP();
+        P_Values = new double[V];
         System.arraycopy(Pk_Values, 0, P_Values, 0, V);
         System.arraycopy(Pk_Values, V, W_Values, 0, 6);
     }
@@ -53,8 +53,8 @@ public class MMV_CalculationFragment extends Model_V_CalculationFragment {
         FragmentManager myFragmentManager = getFragmentManager();
         Fragment fragment = null;
         Bundle args = new Bundle();
-        args.putFloatArray(P_VALUES, P_Values);
-        args.putFloatArray(W_VALUES, W_Values);
+        args.putDoubleArray(P_VALUES, P_Values);
+        args.putDoubleArray(W_VALUES, W_Values);
         //switch (/* Settings choice */) {
         //    case R.id.bar_graph:
         fragment = new DoubleBarGraphFragment();
